+++ lib/strong_migrations/checks.rb
@@ -47,6 +47,11 @@ Then add the NOT NULL constraint in separate migrations."
           code: backfill_code(table, column, default),
           append: append,
           rewrite_blocks: adapter.rewrite_blocks
+      elsif default.respond_to?(:call)
+        # adding a column with a VOLATILE default is not safe
+        # https://www.postgresql.org/docs/current/sql-altertable.html#SQL-ALTERTABLE-NOTES
+        # functions like random() and clock_timestamp() are VOLATILE
+        raise_error :add_column_default_callable
       end
 
       if type.to_s == "json" && postgresql?